Renamed src files to fix autoload issue
After reading about a similar issue (https://stackoverflow.com/questions/30313943/composer-autoload-issue#answer-30455192), I found that renaming the src files to the names of the Classes that should be autoloading is how Composer automatically maps the classes.

Changed the names of the files to proper mixed case, as that seems to be what the autoloader is expecting for PSR-4.