Initial README
As part of the current issue, documentation was needed, so I created the
structure for the **README** file.

I have put the basic information on how to use this code package as well
as the **ApiResponse** class, but there is still documentation needed on
how to use the **GenericFunctions.php** and **HttpPost.php** files.